curve looks good, started motion

+++ Engine3D/Bezier1D.h
@@ -16,7 +16,7 @@ public:
 	IndexedModel GetLine() const;	//generates a model for MeshConstructor Constructor with resT vertices. See Mesh.h/cpp for a IndexedModel defenition and how to intialize it
 	glm::vec4 GetControlPoint(int segment, int indx) const; //returns a control point in the requested segment. indx will be 0,1,2,3, for p0,p1,p2,p3
 	
-	glm::vec4 GetPointOnCurve(int segment, int t) const; //returns point on curve in the requested segment for the value of t
+	glm::vec4 GetPointOnCurve(int segment, float t) const; //returns point on curve in the requested segment for the value of t
 	glm::vec3 GetVelosity(int segment, int t); //returns the derivative of the curve in the requested segment for the value of t
 
 	void AddSegment(glm::vec4 p1, glm::vec4 p2, glm::vec4 p3);  // adds a segment at the end of the curve
